Fixed composition support for D2D 1.0
In order to separate between the lack of support for compositing in D2D 1.0 and the new composition and blending modes in D2D 1.1, the method SetCompositionMode has been added to the render target resource holder. Its default implementation does nothing for D2D 1.0 except for returning true for the only composition modes available: wxCOMPOSITION_DEST and wxCOMPOSITION_OVER. Support for D2D 1.1 is not yet implemented.